Move models into package.
