+++ app/src/main/java/com/helphelp2/android/MainActivity.java
@@ -15,6 +15,8 @@ import com.android.volley.toolbox.Volley;
 import com.google.android.gms.common.ConnectionResult;
 import com.google.android.gms.common.api.GoogleApiClient;
 import com.google.android.gms.location.LocationServices;
+import com.helphelp2.android.models.Place;
+import com.helphelp2.android.models.PlacesResponse;
 import com.squareup.moshi.JsonAdapter;
 import com.squareup.moshi.Moshi;
