+++ app/src/main/java/com/helphelp2/android/PlacesStorageFragment.java
@@ -4,6 +4,8 @@ import android.location.Location;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
 
+import com.helphelp2.android.models.Place;
+
 import java.util.LinkedList;
 import java.util.List;
